+++ packages/services/f-feature-management/CHANGELOG.md
@@ -3,13 +3,19 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.8.0
+------------------------------
+*March 3, 2022*
+
+### Changed
+ - Replace in-situ feature management library with @justeat/feature-management
+
 v0.7.2
 ------------------------------
 *February 21, 2022*
 
 ### Changed 
  - Fix import of config JSON action import for features store module
-
 
 v0.7.1
 ------------------------------